@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.6'
+gem 'rails', '4.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -8,18 +8,6 @@ gem 'rails', '3.2.6'
 gem 'mysql2'
 
 gem 'json'
-
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
-
-  gem 'uglifier', '>= 1.0.3'
-end
 
 gem 'jquery-rails'
 
@@ -46,3 +34,4 @@ gem 'will_paginate', '~> 3.0.5'
 
 gem 'dynamic_form'
 gem 'jquery-ui-rails'
+gem 'protected_attributes'
