@@ -5,4 +5,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-GestionOpticaArena::Application.config.secret_token = '232bb6b9843fa2471157995d122cf3d1a201391f4267c1c8f204eee64527ef3c088ee4d449ad6c38c950747d5c0890d66055a1f76410751d5d673db652a137ef'
+GestionOpticaArena::Application.config.secret_key_base = '232bb6b9843fa2471157995d122cf3d1a201391f4267c1c8f204eee64527ef3c088ee4d449ad6c38c950747d5c0890d66055a1f76410751d5d673db652a137ef'
